symlink only if no exist

+++ travis/install-autotools.sh
@@ -12,8 +12,12 @@ if [ "${TRAVIS_OS_NAME}" = "osx" ] ; then
     which glibtoolize
     glibtool --version
     mkdir -p ${TOP}/bin
-    ln -s `which glibtool` ${TOP}/bin/libtool
-    ln -s `which glibtoolize` ${TOP}/bin/libtoolize
+    if [ ! -f ${TOP}/bin/libtool ] ; then
+        ln -s `which glibtool` ${TOP}/bin/libtool
+    fi
+    if [ ! -f ${TOP}/bin/libtoolize ] ; then
+        ln -s `which glibtoolize` ${TOP}/bin/libtoolize
+    fi
 fi
 
 if [ "${TRAVIS_OS_NAME}" = "linux" ] ; then
